Hermes/requirements.txt
numpy==1.26.3
transformers==4.44.1
pyAgrum==1.17.2
aiozmq
async_timeout
openai
pyyaml
jsonlines
requests
aiohttp
pydantic-settings
gym
beautifulsoup4
langchain
langgraph
backoff
pandas
icecream
pydantic